+++ src/utils/notif-handler.js
@@ -3,6 +3,7 @@ const express = require("express");
 const https = require("https");
 const _ = require("lodash");
 const requireHullMiddleware = require("./require-hull-middleware");
+const handleExtractFactory = require("./extract-handler-factory");
 const Batcher = require("../infra/batcher");
 
 function subscribeFactory(options) {
@@ -36,7 +37,7 @@ function getHandlerName(eventName) {
   return `${model}:${action}`;
 }
 
-function processHandlersFactory(handlers, userHandlerOptions = {}) {
+function processHandlersFactory(handlers, options = {}) {
   const ns = crypto.randomBytes(64).toString("hex");
   return function process(req, res, next) {
     try {
@@ -50,17 +51,17 @@ function processHandlersFactory(handlers, userHandlerOptions = {}) {
       if (messageHandlers && messageHandlers.length > 0) {
         if (message.Subject === "user_report:update") {
           // optionally group user traits
-          if (notification.message && notification.message.user && userHandlerOptions.groupTraits) {
+          if (notification.message && notification.message.user && options.groupTraits) {
             notification.message.user = client.utils.traits.group(notification.message.user);
           }
           // add `matchesFilter` boolean flag
-          notification.message.matchesFilter = helpers.filterNotification(notification.message, userHandlerOptions.segmentFilterSetting || connectorConfig.segmentFilterSetting);
+          notification.message.matchesFilter = helpers.filterNotification(notification.message, options.segmentFilterSetting || connectorConfig.segmentFilterSetting);
           processing.push(Promise.all(messageHandlers.map((handler, i) => {
             return Batcher.getHandler(`${ns}-${eventName}-${i}`, {
               ctx,
               options: {
-                maxSize: userHandlerOptions.maxSize || 100,
-                maxTime: userHandlerOptions.maxTime || 10000
+                maxSize: options.maxSize || 100,
+                maxTime: options.maxTime || 10000
               }
             })
             .setCallback((messages) => {
@@ -95,58 +96,13 @@ function processHandlersFactory(handlers, userHandlerOptions = {}) {
   };
 }
 
-function handleExtractFactory({ handlers, userHandlerOptions }) {
-  return function handleExtract(req, res, next) {
-    if (!req.body || !req.body.url || !req.body.format || !handlers["user:update"]) {
-      return next();
-    }
 
-    const { client, helpers } = req.hull;
+module.exports = function notifHandler({ handlers = {}, onSubscribe, userHandlerOptions, options }) {
+  if (userHandlerOptions) {
+    console.warn("deprecation: userHandlerOptions has been deprecated in favor of options in notifHandler params. This will be a breaking change in 0.14.x");
+  }
 
-    return helpers.handleExtract({
-      body: req.body,
-      batchSize: userHandlerOptions.maxSize || 100,
-      onResponse: () => {
-        res.end("ok");
-      },
-      onError: (err) => {
-        client.logger.error("connector.batch.error", err.stack);
-        res.sendStatus(400);
-      },
-      handler: (users) => {
-        const segmentId = req.query.segment_id || null;
-        if (userHandlerOptions.groupTraits) {
-          users = users.map(u => client.utils.traits.group(u));
-        }
-        const messages = users.map((user) => {
-          const segmentIds = _.compact(_.uniq(_.concat(user.segment_ids || [], [segmentId])));
-          const account = user.account || {};
-          return {
-            user,
-            segments: _.compact(segmentIds.map(id => _.find(req.hull.segments, { id }))),
-            account
-          };
-        });
-
-        // add `matchesFilter` boolean flag
-        messages.map((m) => {
-          if (req.query.source === "connector") {
-            m.matchesFilter = helpers.filterNotification(m, userHandlerOptions.segmentFilterSetting || req.hull.connectorConfig.segmentFilterSetting);
-          } else {
-            m.matchesFilter = true;
-          }
-          return m;
-        });
-        return handlers["user:update"](req.hull, messages);
-      }
-    }).catch((err) => {
-      client.logger.error("connector.batch.error", err.stack || err);
-    });
-  };
-}
-
-
-module.exports = function notifHandler({ handlers = {}, onSubscribe, userHandlerOptions = {} }) {
+  const _options = options || userHandlerOptions || {};
   const _handlers = {};
   const app = express.Router();
 
@@ -166,7 +122,7 @@ module.exports = function notifHandler({ handlers = {}, onSubscribe, userHandler
     addEventHandlers(handlers);
   }
 
-  app.use(handleExtractFactory({ handlers, userHandlerOptions }));
+  app.use(handleExtractFactory({ handlers, options: _options }));
   app.use((req, res, next) => {
     if (!req.hull.message) {
       const e = new Error("Empty Message");
@@ -177,7 +133,7 @@ module.exports = function notifHandler({ handlers = {}, onSubscribe, userHandler
   });
   app.use(requireHullMiddleware());
   app.use(subscribeFactory({ onSubscribe }));
-  app.use(processHandlersFactory(_handlers, userHandlerOptions));
+  app.use(processHandlersFactory(_handlers, _options));
   app.use((req, res) => { res.end("ok"); });
 
   app.addEventHandler = addEventHandler;